extract spatial-analysis.ts
