+++ components/js-eas-sdk-demo/spatial-analysis.ts
@@ -1,0 +1,48 @@
+import booleanContains from "@turf/boolean-contains";
+import { point, polygon } from "@turf/helpers";
+import { AttestationData, ContainmentResult } from "./types";
+import { getLocation } from "./data-utils";
+
+/**
+ * Tests polygon containment for a set of attestations against a location boundary
+ * @param attestations Array of attestations to test
+ * @param locationAttestation The location attestation containing the polygon boundary
+ * @returns Array of containment results for each attestation
+ */
+export function testPolygonContainment(
+  attestations: AttestationData[],
+  locationAttestation: AttestationData
+): ContainmentResult[] {
+  // Extract polygon from location attestation
+  const locationAttestationJson = JSON.parse(locationAttestation.data);
+  const polygonData = getLocation(locationAttestationJson);
+  console.log(`\npolygon data from location attestation\n`, JSON.stringify(polygonData, null, 2));
+
+  // Create turf polygon from the location attestation
+  const turfPolygon = polygon(polygonData.coordinates);
+
+  const containmentResults = attestations.map((attestation, index) => {
+    const attestationJson = JSON.parse(attestation.data);
+    console.log(`\ndecoded attestation ${index + 1} data\n`, JSON.stringify(attestationJson, null, 2));
+
+    const locationData = getLocation(attestationJson);
+    console.log(`\nextracted location from attestation ${index + 1}\n`, JSON.stringify(locationData, null, 2));
+
+    // Create turf point from the attestation location
+    const turfPoint = point(locationData.coordinates);
+
+    // Test containment
+    const isContained = booleanContains(turfPolygon, turfPoint);
+    console.log(`\nattestation ${index + 1} point ${isContained ? 'IS' : 'IS NOT'} contained in polygon\n`);
+
+    return {
+      attestationId: attestation.uid,
+      location: locationData,
+      isContainedInPolygon: isContained
+    };
+  });
+
+  console.log(`\ncontainment test results (${containmentResults.length} total)\n`, JSON.stringify(containmentResults, null, 2));
+
+  return containmentResults;
+}
